+++ gcs/src/gcs_fifo_lite.c
@@ -32,15 +32,15 @@ gcs_fifo_lite_t* gcs_fifo_lite_create (size_t length, size_t item_size)
 
     /* Find real length. It must be power of 2*/
     while (l < length) l = l << 1;
-    if (l > GU_LONG_MAX) return NULL;
-    if (l * item_size > GU_ULONG_MAX) return NULL;
+    if (l             > (ulong) GU_LONG_MAX) return NULL;
+    if (l * item_size > (ulong) GU_ULONG_MAX) return NULL;
 
     ret = GU_CALLOC (1, gcs_fifo_lite_t);
     if (ret) {
         ret->item_size = item_size;
         ret->length    = l;
 	ret->mask      = ret->length - 1;
-	ret->queue      = gu_malloc (ret->length * item_size);
+	ret->queue     = gu_malloc (ret->length * item_size);
 	if (ret->queue) {
 	    gu_mutex_init (&ret->lock,     NULL);
 	    gu_cond_init  (&ret->put_cond, NULL);
